+++ src/models.rs
@@ -1,5 +1,2 @@
 pub mod public;
 pub(crate) mod internal;
-pub(crate) mod domain;
-
-pub use public::*;
