+++ src/repositories/vector_memory_repository.rs
@@ -1,8 +1,8 @@
 use uuid::Uuid;
 use async_trait::async_trait;
 use crate::errors::custom::CustomError;
-use crate::models::internal::MemoryEntry;
-use crate::models::public::MemoryFilters;
+use crate::models::internal::memory_entry::MemoryEntry;
+use crate::models::public::memory_filters::MemoryFilters;
 
 /// Available embedding models and their corresponding vector sizes
 #[derive(Debug, Clone, Copy)]
